+++ src/app/reservation/[id]/page.tsx
@@ -1,13 +1,12 @@
-import Spacer from '@/components/shuttle-detail/components/Spacer';
-import ShuttleImage from '@/components/shuttle-detail/components/ShuttleImage';
-import KakaoMap from '@/components/shuttle-detail/components/KakaoMap';
-import BackButton from '@/components/shuttle-detail/components/BackButton';
 import Footer from '@/components/footer/Footer';
 import { NOTICE_TYPE } from '@/components/notice-section/NoticeSection';
 import NoticeSection from '@/components/notice-section/NoticeSection';
 import { getRoute } from '@/services/shuttleOperation';
-import { ShuttleInfo } from '@/components/shuttle-detail/components/ShuttleInfo';
 import ReservationForm from './components/ReservationForm';
+import EventInfo from '@/components/event/components/EventInfo';
+import EventImage from '@/components/event/components/EventImage';
+import KakaoMap from '@/components/kakao-map/KakaoMap';
+import BackButton from '@/components/buttons/back-button/BackButton';
 
 interface Props {
   params: {
@@ -31,8 +30,8 @@ const Page = async ({ params, searchParams }: Props) => {
   return (
     <main className="relative overflow-y-hidden">
       <BackButton />
-      <ShuttleImage image={route.shuttle.image} />
-      <ShuttleInfo shuttle={route.shuttle} status={route.status} type="ROUTE" />
+      <EventImage image={route.shuttle.image} />
+      <EventInfo shuttle={route.shuttle} status={route.status} type="ROUTE" />
       <KakaoMap
         placeName={route.shuttle.destination.name}
         latitude={route.shuttle.destination.latitude}
@@ -46,7 +45,7 @@ const Page = async ({ params, searchParams }: Props) => {
       <NoticeSection type={NOTICE_TYPE.CANCELLATION_AND_REFUND} />
       <NoticeSection type={NOTICE_TYPE.TERM_AND_CONDITION} />
       <Footer />
-      <Spacer />
+      <div className="h-120" />
     </main>
   );
 };
