--- src/components/shuttle-detail/components/Spacer.tsx
@@ -1,5 +1,0 @@
-const Spacer = () => {
-  return <div className="h-120" />;
-};
-
-export default Spacer;
